+++ demos/23-bezier-point-distance/index.js
@@ -58,9 +58,8 @@
       if (line) {
         // Draw the line connecting the mouse/touch position with the closest curve point.
         draw.line(line.a, line.b, "rgba(192,128,0,0.5)", 1);
-        // pb.fill.circleHandle(line.a, 3.0, "rgb(255,192,0)");
         draw.cross(line.a, 3.0, "rgb(192,128,0)", 1.0);
-        fill.text(t.toFixed(3), line.b.x + 3, line.b.y + 3, { color: "rgb(128,128,0)" });
+        fill.text(t.toFixed(3), line.b.x + 3, line.b.y + 3, { color: "rgb(128,128,0)", fontFamily: "Arial", fontSize: 11 });
       }
     };