+++ src/esm/utils/helpers/BezierPathInteractionHelper.js
@@ -193,8 +193,8 @@ export class BezierPathInteractionHelper {
             // removePathVertexDragStartListeners(this.paths[i], listener);
         }
         this.paths = [];
-        this.pb.canvas.removeEventListener("mouseenter", this._mouseEnterListener);
-        this.pb.canvas.removeEventListener("mouseleave", this._mouseLeaveListener);
+        this.pb.eventCatcher.removeEventListener("mouseenter", this._mouseEnterListener);
+        this.pb.eventCatcher.removeEventListener("mouseleave", this._mouseLeaveListener);
         this._mouseHandler.destroy();
         this._keyHandler.destroy();
     }
@@ -359,12 +359,11 @@ export class BezierPathInteractionHelper {
                 _self._clearMoveEvent();
             }
         };
-        // new AlloyFinger(this.pb.canvas, afProps);
         if (window["createAlloyFinger"]) {
-            return window["createAlloyFinger"](this.pb.eventCatcher ? this.pb.eventCatcher : this.pb.canvas, afProps);
+            return window["createAlloyFinger"](this.pb.eventCatcher ? this.pb.eventCatcher : this.pb.eventCatcher, afProps);
         }
         else {
-            return new AlloyFinger(this.pb.eventCatcher ? this.pb.eventCatcher : this.pb.canvas, afProps);
+            return new AlloyFinger(this.pb.eventCatcher ? this.pb.eventCatcher : this.pb.eventCatcher, afProps);
         }
     }
     // +---------------------------------------------------------------------------------
@@ -372,7 +371,7 @@ export class BezierPathInteractionHelper {
     // +-------------------------------
     _installMouseListener() {
         var _self = this;
-        var mouseHandler = new MouseHandler(this.pb.canvas)
+        var mouseHandler = new MouseHandler(this.pb.eventCatcher)
             .up(function (e) {
             if (e.params.wasDragged)
                 return;
@@ -412,20 +411,11 @@ export class BezierPathInteractionHelper {
             _self.onVertexInserted(_self.currentPathIndex, leftPath.bezierCurves.length, newPath, oldPath);
         })
             .move(function (e) {
-            // console.log('moved');
-            // if( _self.pb.getDraggedElementCount() == 0 )
             _self.mouseIsOver = true;
             _self._handleMoveEvent(e.params.pos.x, e.params.pos.y);
         });
-        // _self.pb.canvas.addEventListener("mouseenter", function () {
-        //   _self.mouseIsOver = true;
-        // });
-        // _self.pb.canvas.addEventListener("mouseleave", function () {
-        //   _self.mouseIsOver = false;
-        //   _self._clearMoveEvent();
-        // });
-        this.pb.canvas.addEventListener("mouseenter", this._mouseEnterListener);
-        this.pb.canvas.addEventListener("mouseleave", this._mouseLeaveListener);
+        this.pb.eventCatcher.addEventListener("mouseenter", this._mouseEnterListener);
+        this.pb.eventCatcher.addEventListener("mouseleave", this._mouseLeaveListener);
         return mouseHandler;
     }
     // +---------------------------------------------------------------------------------